Add universal action, hotkeys as triggers
Adds keyboard hotkeys to trigger the workflow for the current selection
or the clipboard contents. Also adds a universal action trigger if the
selection is text (as opposed to file or URL).

The keyboard shortcuts are not part of the `info.plist` and even if they
were, they would be ignored when imported.

Relates to: #12