Add a GitHub action for release
This adds a GitHub action to build and release the workflow whenever a
tag is pushed that looks like a semver tag.

When a release with a corresponding tag is created through the GitHub
interface, this action runs. It builds and packages the workflow and
then adds the workflow to the existing GitHub release.

If a tag is pushed, but no GitHub release exists, it creates a new one.
The title is going to be the pushed tag and the body the commit message
of the commit the tag points to.

Relates to: #12