fix: set user agent
dict.cc responds with "Please don't run crawlers against dict.cc and
don't try to make the dictionary available offline."

Since we're not doing any of the two, we just be a good citizen and
report who we are. If there's actually an issue with the usage of
dict.cc from this workflow it's easy to block. We can then see if we
come to a solution. Fixes #19